moved tests to appropriate folders #19
